manage delete in customers

+++ abc-recycling-app/src/components/pages/Customers.js
@@ -67,12 +67,22 @@ const Customers = props => {
       name: "",
       button: true,
       width: '60px',
-      cell: row => (
-        <FaTimes
+      cell: row => {
+        if (row.can_delete === "1") {
+          return (
+            <FaTimes
             style={{color: '#D83232', cursor: 'pointer', transform: 'scale(1.5)'}}
             onClick={() => deleteCustomer(row.id)}
-        />
-      )
+            />
+          );
+        } else {
+          return (
+            <FaTimes
+            style={{color: 'grey', cursor: 'pointer', transform: 'scale(1.5)'}}
+            />
+          );
+        }
+      },
     },
   ];
 
@@ -132,6 +142,7 @@ const Customers = props => {
         setCustomerList(
           onlyCompanies ? response.data.filter(e => e.type === 'C') : response.data.filter(e => e.type === 'P') 
         );
+        console.log(customerList)
       }
     )
   }, [onlyCompanies]);
